# Faster topic index Fixes #4820