--- a/Sources/MessageIndex.php
+++ b/Sources/MessageIndex.php
@@ -241,8 +241,8 @@ function MessageIndex()
 
 	// Default sort methods tables.
 		$sort_methods_table = array(
-			'subject' => 'JOIN {db_prefix}messages mf ON ("mf"."id_msg" = "t"."id_first_msg")',
-			'starter' => 'JOIN {db_prefix}messages mf ON ("mf"."id_msg" = "t"."id_first_msg") 
+			'subject' => 'JOIN {db_prefix}messages mf ON (mf.id_msg = t.id_first_msg)',
+			'starter' => 'JOIN {db_prefix}messages mf ON (mf.id_msg = t.id_first_msg) 
 							LEFT JOIN {db_prefix}members AS memf ON (memf.id_member = mf.id_member)',
 			'last_poster' => 'JOIN {db_prefix}messages AS ml ON (ml.id_msg = t.id_last_msg)
 							LEFT JOIN {db_prefix}members AS meml ON (meml.id_member = ml.id_member)',
@@ -358,20 +358,13 @@ function MessageIndex()
 		else
 			$enableParticipation = false;
 
-		$sort_table = 'SELECT
-						"t"."id_topic",
-						"t"."id_first_msg",
-						"t"."id_last_msg"
-					FROM
-						{db_prefix}topics t
-						' . (empty($sort_methods_table[$context['sort_by']]) ? '' : $sort_methods_table[$context['sort_by']]) . '
-					WHERE
-						"t"."id_board" = {int:current_board}
-					ORDER BY
-						' . ($pre_query ? $smcFunc['db_custom_order']('t.id_topic', $topic_ids) : 'is_sticky' . ($fake_ascending ? '' : ' DESC') . ', ' . $_REQUEST['sort'] . ($ascending ? '' : ' DESC')) . ' 
-					LIMIT
-						{int:maxindex}
-					' . ($pre_query ? '' : 'OFFSET {int:start} ')
+		$sort_table = 'SELECT t.id_topic, t.id_first_msg, t.id_last_msg
+					FROM {db_prefix}topics t
+					' . (empty($sort_methods_table[$context['sort_by']]) ? '' : $sort_methods_table[$context['sort_by']]) . '
+					WHERE t.id_board = {int:current_board}
+					ORDER BY ' . ($pre_query ? $smcFunc['db_custom_order']('t.id_topic', $topic_ids) : 'is_sticky' . ($fake_ascending ? '' : ' DESC') . ', ' . $_REQUEST['sort'] . ($ascending ? '' : ' DESC')) . ' 
+					LIMIT {int:maxindex}
+					' . ($pre_query ? '' : ', {int:start} ')
 						;
 		
 		$result = $smcFunc['db_query']('substring', '
@@ -392,8 +385,8 @@ function MessageIndex()
 				' . (!empty($message_index_selects) ? (', ' . implode(', ', $message_index_selects)) : '') . '
 			FROM (' . $sort_table . ') as st 
 				JOIN {db_prefix}topics AS t ON (st.id_topic = t.id_topic)
-				JOIN {db_prefix}messages AS ml ON (ml.id_msg = t.id_last_msg)
-				JOIN {db_prefix}messages AS mf ON (mf.id_msg = t.id_first_msg)
+				JOIN {db_prefix}messages AS ml ON (ml.id_msg = st.id_last_msg)
+				JOIN {db_prefix}messages AS mf ON (mf.id_msg = st.id_first_msg)
 				LEFT JOIN {db_prefix}members AS meml ON (meml.id_member = ml.id_member)
 				LEFT JOIN {db_prefix}members AS memf ON (memf.id_member = mf.id_member)' . (!empty($settings['avatars_on_indexes']) ? '
 				LEFT JOIN {db_prefix}attachments AS af ON (af.id_member = memf.id_member)