This is the "fix" for the issue: https://github.com/SimpleMachines/SMF2.1/issues/4820

jemeter result (1k calls):
default (order by last message id desc)
old 61ms
new 59ms
"subject" order
old 73ms
new 60ms
"started by" order
old 73ms
new 60ms

i did this was i mention in the issue -> added a sub query who do the sort and limit stuff
because this is a subquery our limit 123,13 logic doesn't work so i had to use the generic version limit 123 offset 13 otherwise the stuff will not work with pg

in the core the logic already exists in case of $pre_query logic,
so the most code diff came from removing it the deletion of the tab/space.